test: add debugging logging to visualize data flow
This commit attempts to add some logging calls to see exactly how the
data is flowing from the json to the js to the isml. The blocker here is
that only some of the `context.content` attributes are actually making
it through. You can see that `${pdict.title}` for example always
evaluates to null, even if we pass in a `title` string in the page
designer.