feat: create metadata impot files
This commit creates the necessary `.xml` and `.zip` files so that users
can run the 'site import' feature on SFCC to import imgix custom site
settings. We need these settings in order to pass down an imgix API key
to the page designer. These settings will likely be exapnded upon in a
later commit.