feat: create a search service endpoint
This commit creates the `imgixApi.search` method that allows the
searchbar component to search for assets. This is a placeholder
endpoint, as in a future commit this should get refactored to be cleaner
and less error prone. Ideally, we'd actually hit a `/search` endpoint
rather than use the same valye for tags, filename, etc to filter.