feat: add loading state to asset grid
This commit creates the `LoadingSpinner` component and its corresponding
stylesheets. It also removes the `AssetGridContainer` componenet and
instead moves the applicable logic to the `AssetBrowser.tsx` component.

Before this change, the `ImgixProvider` component was being used to
provide the images with the `domain` prop. This was uneccesary. The
`domain` prop can be passed directly to each `Imgix` grid item
component.

This change also now adds a loading indicator when loading sources and
assets. In the event that no sources or no assets are present, a simple
message is displayed.