refactor(asset-grid): render placeholder
This commit adds `errors` and `loading` as accepted props in order to
handle the placeholder rendering logic inside the component.

This will probably be re-worked in a future commit, as this should
really be handled further up the component strucutre. But, for now, this
way of handling showing the loading spinner, placehilder, and errors is
better than what was being done at the `AssetBrowser.tsx` level.