feat: create asset browsner container
This commit creates the `AssetBrowserContainer` component. This
component houses all the lifted states from `AssetBrowser` as one Class
component state. Each slice of state that is passed as a prop has its
own setter function that is also passed along as a prop.

This component also houses all the api-wrapping functions for fetching
and searching for assets. Each of those functions is also passed along
as a prop to the `AssetGrid` component.

An effort was made to make this into a functional component. However, I
quickly discovered that created more "commits" in the react-dev-tools
profiler. I'm therefore choosing to leave this as is despite the
code-smell of having such a large state in this component.