chore: create and export the cursor type
This commit creates the `CursorT` type. The `cursor` is part of the
`meta` value of the response object to an imgix API requet. The `cursor`
holds pagination information about the assets requested. Whenever the
page or cursor parameters are used to filter the assets, the cursor
object updates to reflect the state of the asset pagination. For
example, it holds the `cursor.next` index value, which typically the
value of `cursor.current` plus the `limit` value passed to the
pagination parameters.