feat: create the pagination component
This commit creates the `Pagination.tsx` component. The component
renders two buttons that invoke the `handlePageChange` function prop
when clicked. One button passes `-1` as an arugment and the other `1`,
IE one decrements and the other incremenets the "page index".

The buttons only render if the `cursor` prop has `totalRecords` that
evaluate to `true`.