refactor: lift asset-browser state into container
This commit removes the use of `React.useState` from the component and
instead lifts the component state into a container component,
`AssetBrowserContainer`. The state-setters are accepted as props to this
component, IE `setQuery`.

The commit also removes the need for rendering the `AssetGrid`
placeholder, and instead passes on that responsability to the
`AssetGrid`.

Finally, the commit creates a `handleSearch` function that invokes the
`setLoading` and `setQuery` props and, if a source is currently
selected, runs `requestAssetsFromSource` prop to search for assets using
the `query` as a filter.