feat: display API warnings in asset browser
This commit adds a `placeholder` prop to the `AssetGrid` component. It
also adds an `errors` state to the `AssetBrowser` component. The
`imgixAPI` error messages are passed to the `placeholder` prop when they
occur to more easily inform the user of any API errors.

The commmit also moves the loading spinner inside the grid. This makes
it easier to pass the spinner as the `placeholder` prop whenever the
`loading` state is `true`.