build: check-in react bundles mvp
This commit adds the React app bundles needed to render the sidebar and
react applications in the cartridges. These are still are work in
progress, and this commit should probably be dropped or the bundles
replaced in a future PR. They are being committed to ensure that we have
a "working" bundle in the next branch at all times.