refactor: move event funcs outside of react app
This commit moves the functions that create event listeners and or
emiters outside of the react application. This fixes an issue where,
whenver state was updated, the event emiters and listeners would stop
working.

Because react re-renders essentially create new DOM nodes, these
functions were losing access to their references nodes. So when we tried
to invoke them again, the listeners no longer existed on the nodes.

By moving these event listeners and emiters outside of the react app, we
ensure that they continue to exist even after a re-render.