chore: import roboto font
This commit changes the default font for the react apps to `Roboto`.
This font is imported when the stylesheet first loads. This is not
ideal, as its not as fast as loading from a local file. However, it gets
around an annoying build issue where the generated font files from `yarn
build` need to moved the the correct SFCC direcotry.

A later commit should probably have this files live somehere in the
cartridge.