refactor: accept selected-source as prop
This commit has `SourceSelect.tsx` accept selected source as a prop.
This way, we no longer need to keep track of it in a local component
state. Moreover, this commit changes the use efffect hook so that it
only updates the selected source if it still exists in the array or
there is no currently selected source.

Lastly, this commit refactor `onClick` into `updateSource`, since
`handleSelect` already updates the `selectedSource` prop.`