feat: style asset on select
This commit adds conditional styling to an asset depending on whether or
not it is selected.

A local `selectedAssetId` keeps track of the currently selected asset.
Whenver an asset is clicked, `onClick` is called updating state and
invoking `handleAssetGridClick`.

The styles are applied to an asset if the asset ID mathces the selected
asset ID.