fix: use the active prop
This commit has the hook make use of its `active` prop. This prop allows
us to enable/disable functionaility depending on whether or not the
element that uses this hook is "active". In our case, when the user is
no longer navigating the dropdown the "Enter" event listener should do
nothing inside this hook.

This fixes a bug where the "Enter" keypress would always cause a source
ID update, regardless of what DOM node triggered the event.