chore(search): refactor handlers
This commit refactors the submit, change, and focus handlers into
fnctions that can be more easily re-used inside this file. Moreover, it
also creates a function, `updateSearchHistory` for the common task of
needing to update the search suggestion list.

This commit also hides the search suggestion list when the input is not
focused. This prevents the suggestion list from being clickable while
it's not acually visisble.