chore: use position property to stack modal
Using position behavior of HTML elements, we can achieve the same effect
of the `z-index` property by modifying the position stack context of
certain elements.

This commit changes the position context of the asset grid items so that
they can be statically positioned, allowing for the source select
dropdown to, when absolutely positioned, render above the asset grid
items.

This change also made it necessary to change the way that filenames were
faded out. Previously, this was done using psuedo elements and
`position: absolute`. Because of the new position context this was no
longer possible.

Instead, this commit makes use of ` linear-gradient` for a
`background-image`. We can then `clip` the `text` against this gradient,
filling the overlap with the color `transparent`.

The drawback to this approach is cross-browser compatibility is
limiteed. But in the absolute worst case, some users will not see faded
out filenames but still see the modal element stacked above the asset
grid items.