fix: add min-height to asset container
This commit adds a `min-height` of `730px` to the `.assetBrowser` class.
This fixes an issue where the container would not take up the entire
height of the modal.

This will mean that on smaller windows/devices the frame will be
scrollable.

Part of the problem is a script SFCC runs in the background that listens
for windoow and or frame resizing. It impedes us from "reactively"
sizing the contianer as the iframe w/h can be changed at any time by the
SFCC background script. Moreover, tring to get the
`document.offsetHeight` or similar techinques for getting the window
size have proven unreliable. Probably at least in part due to that
background SFCC script.