feat: create extension settings page
This commit creates the extension's `options.js/html` files where the
imgix API key will be set/read. The extension settings can be set from
the `chrome://extensions` settings page. It will not open in a new tab
but, instead, it will open in a pop-up. The extension settings are saved
to the `browser.storage` local storage.

The `options.js` file vaidates the imgix API key input on save and
restore. If they key is invalid or errors out for any other reason, the
value in `browser.storage` will not be updated.