chore: limit filename to 18 chars & show full path on hover
This commit ensures the filename for the asset will not bleed outside
the container. If it does, it is truncated to a max of 18 characters.
The fill filename and filepath are shown whenever the asset image or
filename are hovered over. There is an edge case where, on mobile, the
filename will at times be truncated too aggressively. Being that the
vast majority of users will never see the IM on mobile, this is left as
is. Should be tackeld in a later commit.