fix: ensure extension app doesnt affect layout
Before this commit, the extension appp was being added to a table within
a table within the product attributes table. This made it so that the
other table/table rows on the page would expand to whatever size our
product images table became.

After this commit, a new table is added to the page outside of the
product attributes table. This esures layout is not affected as we add
more and more images.

Since what's being submitted is only the custom attribute JSON, this
change does not affect our ability to save the changes to the custom
attribute.