fix: esnure asset-grid doesn't render if domain is null
There's an edge case here that happens when the `requestSources`
function is invoked with an existing source already selected. For some
reason, there's a chance for the domain to be set to undefined if this
function is invoked with an already selected source in state.

This commit ensures the application won't break if there's no domain or
domain is null, but rather will just not render the asset-grid.