fix: locking logic should only factor in product locked state
We no longer need to ensure that the JSON field is empty or the array is
empty before rendering <PreductPageImages> as that will no longer lead
the app to crash (we've handled those issues elsewhere). Furthermore,
the logic as it stood impeded producst without imgix images to be
editble.

This commit instead disables the compononent depending on whether or not
the lock icons is present alone.