feat: add srcset values to product model
This commit adds a `srcSet` value to the product image model. So that,
in an ISML tempalte, you could set and imaage `srcset` value using
`pdict.image.srcSet` , or just `image.srcSet` depending on your tempalte
setup.

Before this commit, the product image model did not automtatically
generate a srcset for the image.

After this commit, the product image model adds a `srcSet` value for
custom attribute images and proxy images.