refactor: create image model data helper functions
This commit refactors the ProductImageModelData modification logic into
several builder functions to simplify the cognitive load of this file.
It also esnures that proxied images have a `srcSeet` attribute added to
their model data that includes all the of the `srcset` generated by
js-core and not just the 3 sizes we anticiapte from the SFRA
architecture.

It's possible that in a future commit, we do away with the hard-coded
sizes for custom attribute images. This depends on what/how we can
expect customers to use `sizes` and `srcset` in the context of SFCC. At
the moment it's not entirely clear if they would rather pick and choose
from a large `srcset` or have our opinionted 3 sizes be the default.